backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pyyaml==6.0.1
requests==2.32.3
gTTS==2.5.1
